test: add 21 unit tests for Estimator, SlidingWindow, StereoMatcher
- Estimator: construction, camera model injection, error paths, bootstrap
- SlidingWindow: add/reject/evict/clear, keyframe pose queries, optimize guards
- StereoMatcher: config defaults, match confidence, empty input handling
- Also auto-fix assert_eq!(x, true) → assert!(x) in existing test code
- Total: 156 tests passing (up from 135)

+++ tests/hotpath_allocation_test.rs
@@ -95,7 +95,7 @@ async fn test_sustained_zero_allocation_processing() {
                 i,
                 left.clone(),
                 right.clone(),
-                i64::from(i) * 1_000_000,
+                i * 1_000_000,
                 None,
             )
             .await;
@@ -111,7 +111,7 @@ async fn test_sustained_zero_allocation_processing() {
                 i,
                 left.clone(),
                 right.clone(),
-                i64::from(i) * 1_000_000,
+                i * 1_000_000,
                 None,
             )
             .await;
@@ -161,7 +161,7 @@ async fn test_concurrent_processing_no_allocation_growth() {
         let l = left.clone();
         let r = right.clone();
         handles.push(tokio::spawn(async move {
-            est.process_frame_async(i, l, r, i64::from(i) * 1_000_000, None)
+            est.process_frame_async(i, l, r, i * 1_000_000, None)
                 .await
         }));
     }
